pyproject.toml
[project]
name = "agentlab"
version = "0.1.0"
description = "Learning platform for LLMs, MCP, and RAG systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.13",
    "langchain-core>=0.3.21",
    "langchain-openai>=0.2.14",
    "langgraph>=0.2.60",
    "langgraph-checkpoint-sqlite>=2.0.5",
    "mysql-connector-python>=9.1.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.1.1",
    "pytest>=8.4.2",
    "pinecone>=7.3.0",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentlab"]

[tool.basedpyright]
exclude = [
    "tests",
    "**/__pycache__"
]

[tool.pytest.ini_options]
addopts = "-v --capture=no"
